# docs/openapi.json
{
  "components": {
    "responses": {
      "Error": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "An error."
      }
    },
    "schemas": {
...
      "Error": {
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "LatestRelease": {
        "properties": {
          "notes": {
            "type": "string"
          },
          "prerelease": {
            "type": "boolean"
          },
          "published_at": {
            "format": "date-time",
            "type": "string"
          },
          "version": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "Server": {
        "properties": {
          "approvals_required": {
            "type": "boolean"
          },
          "latest": {
            "$ref": "#/components/schemas/LatestRelease"
          },
          "latest_error": {
            "type": "string"
          },
          "mode": {
            "type": "string"
          },
          "mode_note": {
            "type": "string"
          },
          "pending_approval_id": {
            "type": "string"
          },
          "stamped": {
            "type": "boolean"
          },
          "state": {
            "$ref": "#/components/schemas/State"
          },
          "update_available": {
            "type": "boolean"
          },
          "version": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "State": {
        "properties": {
          "backup": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "finished_at": {
            "format": "date-time",
            "type": "string"
          },
          "from_version": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "phase": {
            "type": "string"
          },
          "requested_by": {
            "type": "string"
          },
          "started_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          },
          "version": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "session": {
        "description": "A console session. Changes also need the X-CSRF-Token header.",
        "in": "cookie",
        "name": "retune_session",
        "type": "apiKey"
      },
      "token": {
        "description": "An API token, rtk_…",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "The API the Retune console uses. Sign in with a session (the retune_session cookie, with the X-CSRF-Token header on every change) or send an API token as `Authorization: Bearer rtk_…`. Each operation's `x-retune-access` names who may call it. Errors are `{\"code\": \"…\", \"message\": \"…\"}`.",
    "title": "Retune admin API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/server": {
      "get": {
        "description": "Any signed-in admin or API token; a scoped admin sees only their devices.",
        "operationId": "getServer",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Server"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "token": []
          }
        ],
        "summary": "The server's version, the newest verified release, and any update in progress",
        "tags": [
          "Server"
        ],
        "x-retune-access": "scoped"
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/admin/v1"
    }
  ]
}